+++ CHANGELOG.md
@@ -3,5 +3,10 @@
 # kotlin-const-string-interpolation-plugin Changelog
 
 ## [Unreleased]
+
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- Inlay Hint Collector
+- Inlay Hint Provider
+- Plugin configuration
+- Unit tests